requirements.txt
requests == 2.27.1
pyyaml == 6.0
